chore(refactor): merge resolver and schema for each typedef into single files
