--- resolvers/charts.js
@@ -1,9 +1,0 @@
-import predictedSavings from "../dummy-data/predicted-savings.js";
-
-const chartsResolvers = {
-  Query: {
-    charts: () => ({ predictedSavings }),
-  },
-};
-
-export default chartsResolvers;
