--- resolvers/index.js
@@ -1,6 +1,0 @@
-import userResolvers from "./user.js";
-import chartsResolvers from "./charts.js";
-
-const resolvers = { ...userResolvers, ...chartsResolvers };
-
-export default resolvers;
