+++ schema/index.js
@@ -1,6 +1,6 @@
 import { gql } from "apollo-server";
-import { typeDef as User } from "./user.js";
-import { typeDef as Charts } from "./charts.js";
+import { typeDef as User, userResolvers } from "./user.js";
+import { typeDef as Charts, chartsResolvers } from "./charts.js";
 
 const root = gql`
   type Query {
@@ -13,5 +13,6 @@ const root = gql`
 `;
 
 const schemaArray = [root, User, Charts];
+const resolvers = { ...userResolvers, ...chartsResolvers };
 
-export default schemaArray;
+export { schemaArray, resolvers };
