+++ schema/user.js
@@ -15,4 +15,15 @@ const typeDef = gql`
   }
 `;
 
-export { typeDef };
+const userResolvers = {
+  Query: {
+    user: async (root, {}, { accessToken, dataSources }) => {
+      const userSettings = await dataSources.walloraAPI.getUserSettings(
+        accessToken
+      );
+      return { userSettings };
+    },
+  },
+};
+
+export { typeDef, userResolvers };
